rename history transforms to on-history
