+++ lib/transforms/index.js
@@ -127,6 +127,16 @@ import {
 } from './on-selection'
 
 /**
+ * On history.
+ */
+
+import {
+  redo,
+  save,
+  undo,
+} from './on-history'
+
+/**
  * Normalize.
  */
 
@@ -134,16 +144,6 @@ import {
   normalizeDocument,
   normalizeSelection,
 } from './normalize'
-
-/**
- * History.
- */
-
-import {
-  redo,
-  save,
-  undo,
-} from './history'
 
 /**
  * Export.
@@ -269,18 +269,18 @@ export default {
   moveToRangeOf,
 
   /**
-   * Normalize.
-   */
-
-  normalizeDocument,
-  normalizeSelection,
-
-  /**
    * History.
    */
 
   redo,
   save,
   undo,
+
+  /**
+   * Normalize.
+   */
+
+  normalizeDocument,
+  normalizeSelection,
 
 }
