+++ lib/transforms/on-history.js
